+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.11] - 2020-09-14
+### Changed
+- Fixed cloning pages with scheduled tasks
+- Added spacing tool button in editor toolbar
+- Bugfixes
+
 ## [2.3.10] - 2020-08-27
 ### Changed
 - Fixed layer tool save-button